+++ tests/test_manifest.py
@@ -115,7 +115,7 @@ class TestBucketKey:
 class TestLockPackageInstallMode:
     def test_serializes_install_mode_default(self):
         pkg = LockPackage(
-            name="foo/bar", pkg_type="skills-pack", url="https://github.com/foo/bar",
+            name="foo/bar", pkg_type="bundle", url="https://github.com/foo/bar",
             version="1.0.0", commit="abc123", bucket="project",
         )
         d = pkg.to_dict()
@@ -123,7 +123,7 @@ class TestLockPackageInstallMode:
 
     def test_serializes_install_mode_symlink(self):
         pkg = LockPackage(
-            name="foo/bar", pkg_type="skills-pack", url="https://github.com/foo/bar",
+            name="foo/bar", pkg_type="bundle", url="https://github.com/foo/bar",
             version="1.0.0", commit="abc123", bucket="project",
             install_mode="symlink",
         )
@@ -132,7 +132,7 @@ class TestLockPackageInstallMode:
 
     def test_roundtrips_install_mode(self):
         src = LockPackage(
-            name="foo/bar", pkg_type="skills-pack", url="https://github.com/foo/bar",
+            name="foo/bar", pkg_type="bundle", url="https://github.com/foo/bar",
             version="1.0.0", commit="abc123", bucket="project",
             install_mode="symlink",
         )
@@ -141,7 +141,7 @@ class TestLockPackageInstallMode:
 
     def test_reads_legacy_lock_without_install_mode(self):
         legacy = {
-            "type": "skills-pack",
+            "type": "bundle",
             "url": "https://github.com/foo/bar",
             "version": "1.0.0",
             "commit": "abc123",
@@ -154,14 +154,14 @@ class TestLockPackageInstallMode:
     def test_rejects_unknown_install_mode(self):
         with pytest.raises(ValueError, match="install_mode must be one of"):
             LockPackage(
-                name="foo/bar", pkg_type="skills-pack", url="https://x",
+                name="foo/bar", pkg_type="bundle", url="https://x",
                 version="1.0.0", commit="abc", bucket="project",
                 install_mode="hardlink",
             )
 
     def test_to_dict_omits_install_mode_when_default(self):
         pkg = LockPackage(
-            name="foo/bar", pkg_type="skills-pack", url="https://x",
+            name="foo/bar", pkg_type="bundle", url="https://x",
             version="1.0.0", commit="abc", bucket="project",
         )
         d = pkg.to_dict()
@@ -169,7 +169,7 @@ class TestLockPackageInstallMode:
 
     def test_to_dict_includes_install_mode_when_non_default(self):
         pkg = LockPackage(
-            name="foo/bar", pkg_type="skills-pack", url="https://x",
+            name="foo/bar", pkg_type="bundle", url="https://x",
             version="1.0.0", commit="abc", bucket="project",
             install_mode="symlink",
         )